rag_classification/requirements.txt
sentence-transformers>=2.7,<3.0
annoy>=1.17
boto3>=1.34          # only needed for LLM_PROVIDER=bedrock
requests>=2.31       # used by Ollama client
python-dotenv>=1.0
